+++ libft/ft_strlen.c
@@ -15,6 +15,7 @@
 size_t	ft_strlen(const char *s)
 {
 	size_t	i;
+
 	i = 0;
 	while (s && s[i] != '\0')
 	{
