+++ srcs/built_ins/export/export_utils.c
@@ -1,0 +1,76 @@
+#include "../../../includes/mini_shell.h"
+
+extern t_env_var	env;
+
+char	**new_env(char *str)
+{
+	char	**new;
+	int		length;
+	int		i;
+
+	length = 0;
+	i = 0;
+	while (env.env_var[length] != NULL)
+		length++;
+	new = (char **)malloc(sizeof(char *) * (length + 2));
+	while (env.env_var[i] != NULL)
+	{
+		new[i] = ft_strdup(env.env_var[i]);
+		i++;
+	}
+	new[i] = ft_strdup(str);
+	i++;
+	new[i] = NULL;
+	free_2d_array(env.env_var);
+	return (new);
+}
+
+void	export_new_env(char *var, char *tmp)
+{
+	var = ft_strjoin(var, tmp);
+	free(tmp);
+	env.env_var = new_env(var);
+	free(var);
+}
+
+int	op_not(char c)
+{
+	if (c == '=')
+		return (1);
+	else if (c == '+')
+		return (1);
+	else if (c == '/')
+		return (1);
+	else if (c == '*')
+		return (1);
+	else if (c == '-')
+		return (1);
+	else if (c == '%')
+		return (1);
+	return (0);
+}
+
+char	*fetch_var(char *var, int *i)
+{
+	char	*value;
+
+	value = ft_calloc(1, ft_strlen(var) + 1);
+	*i = 0;
+	while (var[*i])
+	{
+		if (var[*i] == '=')
+			break ;
+		value[*i] = var[*i];
+		(*i)++;
+	}
+	value[(*i)] = '\0';
+	return (value);
+}
+
+int	export_error(char *value, char *var)
+{
+	free(value);
+	// printf("%s: not a valid identifier\n", ft_strchr(var, '='));
+	free(var);
+	return (EXIT_FAILURE);
+}
