+++ srcs/executor/executor.c
@@ -105,10 +105,12 @@ char *get_abs_cmd(char *cmd)
         return (ft_strdup(cmd));
     }
     init(&path_split);
+    // ft_putstr_2d(path_split);
     i = 0;
     while (path_split[i])
     {
-        abs_cmd_path = get_abs_cmd_path(&abs_cmd_path,path_split[i], cmd);
+        abs_cmd_path = get_abs_cmd_path(&abs_cmd_path, path_split[i], cmd);
+        // ft_putendl_fd(abs_cmd_path, 2);
         dup = ft_strdup(abs_cmd_path);
         if (abs_cmd_path == NULL)
             return (NULL);
@@ -123,6 +125,8 @@ char *get_abs_cmd(char *cmd)
         {
             free_me(&dup);
             free_me(&abs_cmd_path);
+            if (i == 2)
+                break ;
         }
         i++;
     }
@@ -282,7 +286,7 @@ void replace_quote (t_pars_tokens *pa_tkns, int i)
 {
     int j;
     j = 1;
-	
+
     while (pa_tkns[i].cmd && pa_tkns[i].cmd[j])
     {
         if(pa_tkns[i].cmd && token_contains_quote(pa_tkns[i].cmd[j]))
@@ -352,6 +356,7 @@ int execute_cmd(t_pars_tokens *pa_tokens, int i)
     {
         if (!ft_eco_check(pa_tokens[i].cmd[0]))
             abs_cmd_path = get_abs_cmd(pa_tokens[i].cmd[0]);
+        // ft_putendl_fd(abs_cmd_path, 2);
     }
     if(abs_cmd_path == NULL)
     {
