+++ srcs/main.c
@@ -14,7 +14,7 @@ static int get_input()
 		// ft_putnbr_fd(env.stat_code, 2);
 		// ft_putstr_fd("\n", 2);
 		// c = 11;
-		input = readline("MS SHELL=====>");
+		input = readline("");
 		if (input == NULL)
 		{
 			free_env();
@@ -27,10 +27,11 @@ static int get_input()
 		input_to_tokens(input);
 		if (env.trigger)
 		{
-			// ft_putnbr_fd(env.stat_code, 1);
+			// ft_putnbr_fd(env.stat_code, 2);
 			// ft_putchar_fd('\n', 1);
 			// printf("%d\n", env.stat_code);
 			exit(env.stat_code);
+			// printf("check\n");
 		}
 		free(input);
 	}
